fix sorting so / is less than all other path characters
File "a-b" sorts before file "a/c" as strings, but "a/c" comes
first in an fs.WalkDir enumeration. So we were incorrectly
displaying D and N in some cases when a new file is added
in a subdirectory. This didn't corrupt the output .lsr itself, only
the commentary. But let's fix it for cleanliness.